requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-appconf==1.0.4
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-decouple==2.1
django-extensions==3.1.3
djangorestframework==3.12.4
drf-jwt==1.19.1
ecdsa==0.17.0
filelock==3.0.12
huggingface-hub==0.0.8
humanize==3.9.0
idna==2.10
joblib==1.0.1
mysqlclient==2.0.3
nltk==3.6.2
nodejs==0.1.1
npm==0.1.1
numpy==1.20.3
oauthlib==3.1.1
optional-django==0.1.0
packaging==20.9
Pillow==8.2.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.17.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
qrcode==6.1
rcssmin==1.0.6
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7.2
sacremoses==0.0.45
scipy==1.7.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tokenizers==0.10.3
tqdm==4.61.0
transformers==4.6.1
urllib3==1.26.5
